+++ src/main/java/repository/implementation/mapper/AdvisorResultSetMapper.java
@@ -14,7 +14,7 @@ public class AdvisorResultSetMapper {
         advisor.setLastName(rs.getString("last_name"));
         advisor.setEmail(rs.getString("email"));
         advisor.setSpeciality(rs.getString("speciality"));
-        advisor.setWorkloadScore(rs.getDouble("workload_score"));
+        advisor.setWorkloadScore(rs.getInt("workload_score"));
         return advisor;
     }
 }
